# openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Axis Platform API",
    "version": "v1"
  },
  "paths": {
...
    "/api/business-object-definitions": {
      "get": {
        "tags": [
          "Business Objects"
        ],
        "summary": "List business object definitions for the current workspace",
        "operationId": "ListBusinessObjectDefinitions",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessObjectDefinitionListItemDtoPagedResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "Business Objects"
        ],
        "summary": "Create an unpublished business object definition",
        "operationId": "CreateBusinessObjectDefinition",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBusinessObjectDefinitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessObjectDefinitionDetailDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/business-object-definitions/{id}": {
      "get": {
        "tags": [
          "Business Objects"
        ],
        "summary": "Get a business object definition",
        "operationId": "GetBusinessObjectDefinition",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessObjectDefinitionDetailDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/business-object-definitions/{id}/unpublished": {
      "put": {
        "tags": [
          "Business Objects"
        ],
        "summary": "Save an unpublished business object definition",
        "operationId": "SaveUnpublishedBusinessObjectDefinition",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveUnpublishedBusinessObjectDefinitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessObjectDefinitionDetailDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/business-object-definitions/{id}/publish": {
      "post": {
        "tags": [
          "Business Objects"
        ],
        "summary": "Publish an unpublished business object definition",
        "operationId": "PublishBusinessObjectDefinition",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishBusinessObjectDefinitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessObjectDefinitionDetailDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BusinessObjectChoiceFieldConfigurationDto": {
        "type": "object",
        "properties": {
          "selectionMode": {
            "$ref": "#/components/schemas/BusinessObjectChoiceSelectionMode"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BusinessObjectChoiceOptionDto"
            }
          }
        },
        "additionalProperties": false
      },
      "BusinessObjectChoiceFieldConfigurationInput": {
        "type": "object",
        "properties": {
          "selectionMode": {
            "$ref": "#/components/schemas/BusinessObjectChoiceSelectionMode"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BusinessObjectChoiceOptionInput"
            }
          }
        },
        "additionalProperties": false
      },
      "BusinessObjectChoiceOptionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "optionKey": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "BusinessObjectChoiceOptionInput": {
        "type": "object",
        "properties": {
          "optionKey": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BusinessObjectChoiceSelectionMode": {
        "enum": [
          "Single",
          "Multiple"
        ],
        "type": "string"
      },
      "BusinessObjectDefinitionDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "objectKey": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/BusinessObjectDefinitionStatus"
          },
          "revision": {
            "type": "integer",
            "format": "int32"
          },
          "latestPublishedVersionNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BusinessObjectFieldDefinitionDto"
            }
          },
          "latestPublishedVersion": {
            "$ref": "#/components/schemas/BusinessObjectDefinitionVersionDto"
          }
        },
        "additionalProperties": false
      },
      "BusinessObjectDefinitionListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "objectKey": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/BusinessObjectDefinitionStatus"
          },
          "revision": {
            "type": "integer",
            "format": "int32"
          },
          "latestPublishedVersionNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "BusinessObjectDefinitionListItemDtoPagedResult": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BusinessObjectDefinitionListItemDto"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "BusinessObjectDefinitionStatus": {
        "enum": [
          "Unpublished",
          "Published"
        ],
        "type": "string"
      },
      "BusinessObjectDefinitionVersionChoiceFieldConfigurationDto": {
        "type": "object",
        "properties": {
          "selectionMode": {
            "$ref": "#/components/schemas/BusinessObjectChoiceSelectionMode"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BusinessObjectDefinitionVersionChoiceOptionDto"
            }
          }
        },
        "additionalProperties": false
      },
      "BusinessObjectDefinitionVersionChoiceOptionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sourceChoiceOptionId": {
            "type": "string",
            "format": "uuid"
          },
          "optionKey": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "BusinessObjectDefinitionVersionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sourceDefinitionId": {
            "type": "string",
            "format": "uuid"
          },
          "versionNumber": {
            "type": "integer",
            "format": "int32"
          },
          "publishedByUserId": {
            "type": "string",
            "format": "uuid"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BusinessObjectDefinitionVersionFieldDto"
            }
          }
        },
        "additionalProperties": false
      },
      "BusinessObjectDefinitionVersionFieldDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sourceFieldDefinitionId": {
            "type": "string",
            "format": "uuid"
          },
          "fieldKey": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "fieldType": {
            "$ref": "#/components/schemas/BusinessObjectFieldType"
          },
          "choiceConfiguration": {
            "$ref": "#/components/schemas/BusinessObjectDefinitionVersionChoiceFieldConfigurationDto"
          },
          "rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BusinessObjectDefinitionVersionFieldRuleDto"
            }
          }
        },
        "additionalProperties": false
      },
      "BusinessObjectDefinitionVersionFieldRuleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sourceFieldRuleId": {
            "type": "string",
            "format": "uuid"
          },
          "bindingId": {
            "type": "string",
            "format": "uuid"
          },
          "bindingRevision": {
            "type": "integer",
            "format": "int32"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "BusinessObjectFieldDefinitionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fieldKey": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "fieldType": {
            "$ref": "#/components/schemas/BusinessObjectFieldType"
          },
          "choiceConfiguration": {
            "$ref": "#/components/schemas/BusinessObjectChoiceFieldConfigurationDto"
          },
          "rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BusinessObjectFieldRuleDto"
            }
          }
        },
        "additionalProperties": false
      },
      "BusinessObjectFieldDefinitionInput": {
        "type": "object",
        "properties": {
          "fieldKey": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "fieldType": {
            "$ref": "#/components/schemas/BusinessObjectFieldType"
          },
          "rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BusinessObjectFieldRuleInput"
            },
            "nullable": true
          },
          "choiceConfiguration": {
            "$ref": "#/components/schemas/BusinessObjectChoiceFieldConfigurationInput"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BusinessObjectFieldRuleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "bindingId": {
            "type": "string",
            "format": "uuid"
          },
          "bindingRevision": {
            "type": "integer",
            "format": "int32"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "BusinessObjectFieldRuleInput": {
        "type": "object",
        "properties": {
          "bindingId": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BusinessObjectFieldType": {
        "enum": [
          "Text",
          "Integer",
          "Decimal",
          "Date",
          "DateTime",
          "Boolean",
          "Choice"
        ],
        "type": "string"
      },
...
      "CreateBusinessObjectDefinitionRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable problem code for client behavior and localization.",
            "nullable": true
          },
          "errorCodes": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Optional field-level machine-readable validation codes keyed by JSON field name.",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "PublishBusinessObjectDefinitionRequest": {
        "type": "object",
        "properties": {
          "expectedRevision": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "SaveUnpublishedBusinessObjectDefinitionRequest": {
        "type": "object",
        "properties": {
          "expectedRevision": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BusinessObjectFieldDefinitionInput"
            }
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}